% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdply.R
\name{fdply}
\alias{fdply}
\title{Read the whole file into a data.table}
\usage{
fdply(input, key.sep = "\\t", sep = "\\t", skip = 0, header = TRUE,
  nblocks = 1, stringsAsFactors = FALSE, select = NULL,
  drop = NULL, col.names = NULL, index = NULL,
  max.size = 536870912, parallel = 1)
}
\arguments{
\item{input}{Path of the input file.}

\item{key.sep}{The character that delimits the first field from the rest.}

\item{sep}{The field delimiter.}

\item{skip}{Number of lines to skip at the beginning of the file}

\item{header}{Whether the file has a header.}

\item{nblocks}{The number of blocks to read. A block is different from a chunk in that chunks can contain many blocks.}

\item{stringsAsFactors}{Whether to convert strings into factors.}

\item{select}{The columns to be read.}

\item{drop}{The columns not to be read.}

\item{col.names}{Names of the columns.}

\item{index}{The index to assign the data.table when it is read.}

\item{max.size}{The maximum size of the chunk (NB: a chunk can contain many blocks).}

\item{parallel}{Number of cores to use.}
}
\value{
A data.table. It is not advised to read a high number of blocks in this way.
}
\description{
Read the whole file into a data.table
}
