% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffply.R
\name{ffply}
\alias{ffply}
\title{Read, process and write.}
\usage{
ffply(input, output = "", FUN, ..., key.sep = "\\t", sep = "\\t",
  skip = 0, header = TRUE, nchunks = Inf, stringsAsFactors = FALSE,
  select = NULL, drop = NULL, col.names = NULL, index = NULL,
  max.size = 536870912, parallel = 1)
}
\arguments{
\item{input}{Path of the input file.}

\item{output}{Path of the output file.}
}
\value{
Returns the number of chunks that were processed. As a side effect,
    writes the processed data.table to the output file.#'
}
\description{
Read, process and write.
}
