% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flply.R
\name{flply}
\alias{flply}
\title{Read, process and return a list.}
\usage{
flply(input, FUN, ..., key.sep = "\\t", sep = "\\t", skip = 0,
  header = TRUE, nchunks = Inf, stringsAsFactors = FALSE,
  select = NULL, drop = NULL, col.names = NULL, index = NULL,
  max.size = 536870912, parallel = 1)
}
\arguments{
\item{input}{Path of the input file.}
}
\value{
Returns a list containing, for each chunk, the result of the
processing.
}
\description{
Read, process and return a list.
}
