% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplyr.R
\docType{package}
\name{fplyr}
\alias{fplyr}
\alias{fplyr-package}
\title{fplyr: Read, Process and Write}
\description{
This package provides a set of functions to quickly read files chunk by
chunk, apply a function to each chunk, and return the result. It is
especially useful when the files to be processed don't fit into the
RAM.
}
\section{Main functions}{

The main functions are ffply and flply. The former writes the processed
data into a file, while the latter returns it as a list. There is also
fdply, which returns a data.table and is useful to only read a certain number
of chunks from the file (one by default).
}

